% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corModels.R
\name{corModelExponential}
\alias{corModelExponential}
\alias{corModelExpRadon2}
\alias{corModelGaussian}
\alias{corModelCauchy}
\alias{corModelSpherical}
\alias{corModelCubic}
\title{Spatial Correlation Models}
\usage{
corModelExponential(distance.matrix)

corModelExpRadon2(distance.matrix)

corModelGaussian(distance.matrix)

corModelCauchy(distance.matrix, extrap)

corModelSpherical(distance.matrix)

corModelCubic(distance.matrix)
}
\arguments{
\item{distance.matrix}{The distance matrix for sampled sites, often divided by the range parameter.}

\item{extrap}{Smoothness parameter (optional)}
}
\value{
Correlation Matrix
}
\description{
Spatial Correlation Models
}
\section{Functions}{
\itemize{
\item \code{corModelExpRadon2}: Exponential Radon 2 Correlation Structure

\item \code{corModelGaussian}: Gaussian Correlation Structure

\item \code{corModelCauchy}: Cauchy Correlation Structure

\item \code{corModelSpherical}: Spherical Correlation Structure

\item \code{corModelCubic}: Cubic Correlation Structure
}}

