% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estcovparm.R
\name{estcovparm}
\alias{estcovparm}
\title{Estimate Covariance Parameters}
\usage{
estcovparm(response, designmatrix, xcoordsvec, ycoordsvec,
  CorModel = "Exponential", estmethod = "REML", covestimates = c(NA,
  NA, NA), pivec = NA, Vnn = NA)
}
\arguments{
\item{response}{a vector of a response variable, possibly with
missing values.}

\item{designmatrix}{is the matrix of covariates used to regress
the response on.}

\item{xcoordsvec}{is a vector of x coordinates}

\item{ycoordsvec}{is a vector of y coordinates}

\item{CorModel}{is the covariance structure. By default, \code{covstruct} is}

\item{estmethod}{is either the default \code{"REML"} for restricted
maximum likelihood to estimate the covariance parameters and
regression coefficients or \code{"ML"} to estimate the covariance
parameters and regression coefficients.
Exponential but other options include the Spherical and the Gaussian.}

\item{covestimates}{is an optional vector of covariance parameter estimates (nugget, partial sill, range). If these are given and \code{estmethod = "None"}, the the provided vector are treated as the estimators to create the covariance structure.}

\item{pivec}{is a vector of estimated detection probabilities
on each of the sites}

\item{Vnn}{is the covariance matrix for the estimated detection
probabilities}
}
\value{
a list with \itemize{
   \item a vector of estimated covariance parameters
   \item the estimated covariance matrix for all sites
   \item the QR decomposition
   \item a vector of estimated fixed effects
   \item other components used in \code{slmfit}
}
}
\description{
Used to estimate spatial covariance parameters for a few different spatial models, including the Exponential, Gaussian, and Spherical.
Estimated parameters can then be used in \code{predict} to predict unobserved values.
}
\details{
The function is used internally in \code{slmfit}.
}
