% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPBKoutput.R
\name{FPBKoutput}
\alias{FPBKoutput}
\title{Create maps and summaries from FPBK results.}
\usage{
FPBKoutput(pred.info, conf.level = 0.95)
}
\arguments{
\item{pred.info}{is the output from \code{FPBK.pred} in this package.}

\item{conf.level}{is the desired confidence level for the prediction}
}
\value{
maps of the kriged and observed counts, the fitted model variogram,
and any requested normal-based confidence intervals.
}
\description{
The main purpose of this function is to take the results from FPBK and make
readable maps, the fitted variogram model, and brief summaries.
}
