% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detection.R
\name{get_detection}
\alias{get_detection}
\title{Obtains estimates for detection parameters through logistic regression.}
\usage{
get_detection(formula, data, varmethod = "Bootstrap")
}
\arguments{
\item{formula}{is an R linear model formula specifying binary detection as the
response variable as well as covariates associated with detection.}

\item{data}{is the data set with the response column of detection and the covariates to
be used in the logistic regression, and the spatial coordinates for all of the sites.}

\item{varmethod}{is the method to obtain the variance matrix for the detection probabilities, either "Delta" or "Bootstrap".}
}
\value{
a list with \itemize{
  \item Fitted detection coefficients
  \item The asymptotic variance-covariance matrix of the 
  fitted regression coefficients
  \item the formula used to fit the logistic regression model
  \item A list of bootstrapped regression coefficients used to 
  obtain a covariance matrix for detection estimates
  \item The method used to get the variance covariance matrix
  for detection probabilities.
}
}
\description{
Estimates logistic regression coefficients from detection data.
}
\examples{
df <- data.frame(y = rbinom(30, 1, 0.7), x = rnorm(30, 4, 0.4))
get_detection(formula = y ~ x, data = df)
}
