% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reportdoc.R
\name{get_reportdoc}
\alias{get_reportdoc}
\title{Function used to generate an HTML report.}
\usage{
get_reportdoc(output_info)
}
\arguments{
\item{output_info}{is the output from \code{FPBKoutput} in this package.}
}
\description{
This function produces an HTML report with the results from 
the finite population block kriging from \code{FPBKoutput}.
}
