% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MarchMoose}
\alias{MarchMoose}
\title{Data Set with Togiak March Moose Observatations.}
\format{A data frame object:
\describe{
  \item{centrlat}{The latitude of the centroid for each site}
  \item{centrlon}{The latitude of the centroid for each site}
  \item{Strat}{A stratification variable}
  ...
}}
\usage{
MarchMoose
}
\description{
A data set that can be used with the \code{FPBKPack2} package. In
this example, the counts are of moose on 860 sites of equal area.
}
\keyword{datasets}
