% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MarchSight}
\alias{MarchSight}
\title{Simulated Data Set with Moose Radiocollar sightability observations.}
\format{A data frame object including:
\describe{
  \item{Moo.ID}{ID number for the radiocollared moose}
  \item{Detected.}{Y if detected, N if not}
  ...
}}
\usage{
MarchSight
}
\description{
A data set that can be used with the \code{FPBKPack2} package. In
this example, the there are two predictors for detection.
}
\keyword{datasets}
