% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slmfit.R
\name{slmfit}
\alias{slmfit}
\title{Fits a Spatial Linear Model}
\usage{
slmfit(formula, data, xcoordcol, ycoordcol, CorModel = "Exponential",
  estmethod = "REML", covestimates = c(NA, NA, NA),
  detectionobj = NULL, coordtype = "TM", areacol = NULL)
}
\arguments{
\item{formula}{is an R linear model formula specifying density as the
response variable as well as covariates for predicting densities on the unsampled sites.}

\item{data}{is the data set with the response column of densities, the covariates to
be used for the block kriging, and the spatial coordinates for all of the sites.}

\item{xcoordcol}{is the name of the column in the data frame with x coordinates or longitudinal coordinates}

\item{ycoordcol}{is the name of the column in the data frame with y coordinates or latitudinal coordinates}

\item{CorModel}{is the covariance structure. By default, \code{CorModel} is
Exponential but other options include the Spherical and Gaussian.}

\item{estmethod}{is either the default \code{"REML"} for restricted
maximum likelihood to estimate the covariance parameters and
regression coefficients or \code{"ML"} to estimate the covariance
parameters and regression coefficients.}

\item{covestimates}{is an optional vector of covariance parameter estimates (nugget, partial sill, range). If these are given and \code{estmethod = "None"}, the the provided vector are treated as the estimators to create the covariance structure.}

\item{detectionobj}{is a fitted model obj from \code{get_detection}. The default is for this object to be \code{NULL}, resulting in
spatial prediction that assumes perfect detection.}

\item{coordtype}{specifies whether the coordinates are in 
Lat/Lon or already in TM or UTM}

\item{areacol}{is the name of the column with the areas of the sites. By default, we assume that all sites have equal area, in which
case a vector of 1's is used as the areas.}
}
\value{
a list with \itemize{
  \item the spatial covariance estimates
  \item the regression coefficient estimates
  \item the estimated covariance matrix
  \item minus two times the log-likelihood
  \item the names of the predictors used as covariates
  \item the sample size
  \item the name of the correlation model used
  \item a vector of residuals
  \item the design matrix for the sampled sites
  \item a vector of the density of the observed counts
  \item an indicator for whether the user input detection data
  \item a list with information for \code{predict} containing \enumerate{
       \item formula
       \item data, the data set without any missing values
       \item a vector of x-coordinates (in TM)
       \item a vector of y-coordinates (in TM)
       \item whether REML or ML was used
       \item the correlation model used
       \item the covariance matrix for all of the sites
       \item Inverted covariance matrix on the sampled sites
       \item Other items used in \code{predict}
       }
}
}
\description{
Estimates regression coefficients and spatial autocorrelation
parameters.
}
\examples{
slmfit(formula = Moose ~ CountPred + Stratum, data = vignettecount,
xcoordcol = "Xcoords", ycoordcol = "Ycoords")
}
