% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.slmfit.R
\name{summary.slmfit}
\alias{summary.slmfit}
\title{Summarizes a fitted spatial linear model}
\usage{
\method{summary}{slmfit}(object, ...)
}
\arguments{
\item{object}{is an object generated from \code{slmfit}.}

\item{...}{are options to be passed to \code{print}}
}
\value{
a list with \itemize{
  \item model formula
  \item a table of fixed effects estimates and associated standard errors. For models with detection, these are the fixed effects estimates for the abundance model. Due to the complexity in estimating the fixed effects, standard errors and p-values are not given in the summary output when using detection from radiocollar data.
  \item estimated spatial covariance parameter estimates
  \item residuals
       }
}
\description{
Summarizes a fitted spatial linear model
}
