% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignettedetection}
\alias{vignettedetection}
\title{Simulated Data Set with Moose Radiocollar sightability observations.}
\format{A data frame object including:
\describe{
  \item{X}{ID number for the radiocollared moose}
  \item{Detected}{1 if detected, 0 if not}
  \item{DetPred1}{A predictor for detection}
  \item{DetPred2}{A second predictor for detection}
  ...
}}
\usage{
vignettedetection
}
\description{
A data set that can be used with the \code{FPBKPack2} package. In
this example, the there are two predictors for detection.
}
\keyword{datasets}
