% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estcovparm.R
\name{estcovparm}
\alias{estcovparm}
\title{Estimate Covariance Parameters}
\usage{
estcovparm(response, designmatrix, xcoordsvec, ycoordsvec,
  CorModel = "Exponential")
}
\arguments{
\item{response}{a vector of a response variable, possibly with
missing values.}

\item{designmatrix}{is the matrix of covariates used to regress
the response on.}

\item{xcoordsvec}{is a vector of x coordinates}

\item{ycoordsvec}{is a vector of y coordinates}

\item{CorModel}{is the covariance structure. By default, `covstruct` is
Exponential but other options include the Matern, Spherical, and Gaussian.}
}
\value{
a list with a vector of estimated covariance parameters, the fitted
covariance matrix, and the inverse of the fitted covariance matrix.
}
\description{
Used to estimate spatial covariance parameters for a few different spatial models.
Estimated parameters can then be used in `FPBKpred` to predict unobserved values.
}
\details{
The parameters used in the function can be inherited from the FPBKpred function
}
