% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estcovparm.R
\name{estcovparm}
\alias{estcovparm}
\title{Estimate Covariance Parameters}
\usage{
estcovparm(formula, data, xcoordcol, ycoordcol, CorModel = "Exponential")
}
\arguments{
\item{formula}{is  an R linear model formula specifying the counts as the
response variable as well as covariates for predicting counts of animals or plants
on the unsampled sites.}

\item{data}{is the data set with the response column of counts, the covariates to
be used for the block kriging, and the spatial coordinates for all of the sites.}

\item{xcoordcol}{is the name of the column in the data frame with x coordinates}

\item{ycoordcol}{is the name of the column in the data frame with y coordinates}

\item{CorModel}{is the covariance structure. By default, `covstruct` is
Exponential but other options include the Matern, Spherical, and Gaussian.}
}
\value{
a list with a vector of estimated covariance parameters, the fitted
covariance matrix, and the inverse of the fitted covariance matrix.
}
\description{
Used to estimate spatial covariance parameters for a few different spatial models.
Estimated parameters can then be used in `FPBKpred` to predict unobserved values.
}
\details{
The parameters used in the function can be inherited from the FPBKpred function
}
