% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPBKpred.R
\name{FPBKpred}
\alias{FPBKpred}
\title{Perform Finite Population Block Kriging}
\usage{
FPBKpred(formula, data, xcoordcol, ycoordcol, CorModel = "Exponential",
  FPBKcol = NULL, areacol = NULL, detectionest = c(1, 0),
  coordtype = "LatLon")
}
\arguments{
\item{formula}{is an R linear model formula specifying the counts as the
response variable as well as covariates for predicting counts of animals or plants
on the unsampled sites.}

\item{data}{is the data set with the response column of counts, the covariates to
be used for the block kriging, and the spatial coordinates for all of the sites.}

\item{xcoordcol}{is the name of the column in the data frame with x coordinates or longitudinal coordinates}

\item{ycoordcol}{is the name of the column in the data frame with y coordinates or latitudinal coordinates}

\item{CorModel}{is the covariance structure. By default, `CorModel` is
Exponential but other options include the Matern, Spherical, and Gaussian.}

\item{FPBKcol}{is a vector in the data set that contains the weights for
prediction. The default setting predicts the population total}

\item{areacol}{is the name of the column that has the area
for each site of interest. If left blank, it is assumed
that each site has an equal area.}

\item{detectionest}{is a vector of an overall sightability estimate (between `0` and `1`) with a standard error. If the user does not have a standard error, then putting `0` as the standard error will create a confidence interval that is too narrow but will give an accurate prediction for the total.}

\item{coordtype}{specifies whether spatial coordinates are in latitude, longitude (\code{LatLon}) form or UTM (\code{UTM}) form.}
}
\value{
a list with the estimated population total, the estimated prediction
variance, and the vector of predicted counts for all of the sites.
}
\description{
Takes sample data and uses FPBK to predict the counts on the unsampled sites.
The column with the counts should have numeric values for the observed counts
on the sampled sites and `NA` for any site that was not sampled.
}
