% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignettecount}
\alias{vignettecount}
\title{Simulated Data Set with Moose count observations.}
\format{A data frame object including:
\describe{
  \item{X}{ID number for the sites}
  \item{Xcoords}{The x coordinates of the centroid for each site}
  \item{Ycoords}{The y coordinates of the centroid for each site}
  \item{Moose}{The total moose count on each site}
  ...
}}
\usage{
vignettecount
}
\description{
A data set that can be used with the \code{FPBKPack2} package. In
this example, the counts are of moose on 860 sites of equal area.
}
\keyword{datasets}
